Improve text selection and add to text box
