+++ pygame_gui/elements/ui_text_box.py
@@ -1,12 +1,19 @@
 import warnings
 import math
 import html
+import re
 
 from typing import Union, Tuple, Dict, Optional, Any
 
 import pygame
 
+from pygame import MOUSEBUTTONDOWN, MOUSEBUTTONUP, BUTTON_LEFT, KEYDOWN
+from pygame import KMOD_SHIFT, KMOD_META, KMOD_CTRL, KMOD_ALT, K_a, K_c
+from pygame import K_LEFT, K_RIGHT, K_UP, K_DOWN, K_HOME, K_END
+from pygame.event import Event
+
 from pygame_gui.core import ObjectID
+from pygame_gui.core.utility import clipboard_copy
 from pygame_gui._constants import UI_TEXT_BOX_LINK_CLICKED, OldType, UITextEffectType
 from pygame_gui._constants import TEXT_EFFECT_TYPING_APPEAR, TEXT_EFFECT_TILT
 from pygame_gui._constants import TEXT_EFFECT_FADE_IN, TEXT_EFFECT_FADE_OUT, TEXT_EFFECT_BOUNCE
@@ -111,6 +118,8 @@ class UITextBox(UIElement, IUITextOwnerInterface):
                          object_id=object_id,
                          element_id=['text_box'])
         self.should_html_unescape_input_text = should_html_unescape_input_text
+        self.html_text = ""
+        self._plain_text = ""
         if self.should_html_unescape_input_text:
             self.html_text = html.unescape(html_text)
         else:
@@ -139,6 +148,7 @@ class UITextBox(UIElement, IUITextOwnerInterface):
         self.border_colour = None
         self.line_spacing = 1.25
         self.text_cursor_colour = pygame.Color("white")
+        self.selected_bg_colour = pygame.Color(128, 128, 200, 255)
 
         self.link_normal_colour = None
         self.link_hover_colour = None
@@ -166,7 +176,37 @@ class UITextBox(UIElement, IUITextOwnerInterface):
 
         self.parser = None  # type: Optional[HTMLParser]
 
+        self.double_click_timer = self.ui_manager.get_double_click_time() + 1.0
+
+        self.edit_position = 0
+        self._select_range = [0, 0]
+        self.selection_in_progress = False
+        self.cursor_has_moved_recently = False
+
+        self.vertical_cursor_movement = False
+        self.last_horiz_cursor_index = 0
+
         self.rebuild_from_changed_theme_data()
+
+    @property
+    def select_range(self):
+        """
+        The selected range for this text. A tuple containing the start
+        and end indexes of the current selection.
+
+        Made into a property to keep it synchronised with the underlying drawable shape's
+        representation.
+        """
+        return self._select_range
+
+    @select_range.setter
+    def select_range(self, value):
+        self._select_range = value
+        start_select = min(self._select_range[0], self._select_range[1])
+        end_select = max(self._select_range[0], self._select_range[1])
+
+        self.text_box_layout.set_text_selection(start_select, end_select)
+        self.redraw_from_text_block()
 
     def kill(self):
         """
@@ -390,6 +430,54 @@ class UITextBox(UIElement, IUITextOwnerInterface):
         super().update(time_delta)
         if not self.alive():
             return
+
+        scaled_mouse_pos = self.ui_manager.get_mouse_position()
+        if self.hovered:
+            if (self.scroll_bar is None or
+                    (self.scroll_bar is not None and
+                     not self.scroll_bar.hover_point(scaled_mouse_pos[0], scaled_mouse_pos[1]))):
+                self.ui_manager.set_text_hovered(True)
+
+        if self.double_click_timer < self.ui_manager.get_double_click_time():
+            self.double_click_timer += time_delta
+
+        if self.selection_in_progress:
+            scaled_mouse_pos = self.ui_manager.get_mouse_position()
+            text_layout_space_pos = self._calculate_text_space_pos(scaled_mouse_pos)
+            select_end_pos = self.text_box_layout.find_cursor_position_from_click_pos(text_layout_space_pos)
+            new_range = [self.select_range[0], select_end_pos]
+            if new_range[0] != self.select_range[0] or new_range[1] != self.select_range[1]:
+                self.select_range = [new_range[0], new_range[1]]
+
+                self.edit_position = self.select_range[1]
+                self.cursor_has_moved_recently = True
+
+        if self.cursor_has_moved_recently:
+            self.cursor_has_moved_recently = False
+
+            self.text_box_layout.set_cursor_position(self.edit_position)
+            if not self.vertical_cursor_movement:
+                if self.text_box_layout.cursor_text_row is not None:
+                    self.last_horiz_cursor_index = self.text_box_layout.cursor_text_row.get_cursor_index()
+            self.vertical_cursor_movement = False
+            self._handle_cursor_visibility()
+            if self.scroll_bar is not None:
+                cursor_y_pos_top, cursor_y_pos_bottom = self.text_box_layout.get_cursor_y_pos()
+                current_height_adjustment = int(self.scroll_bar.start_percentage *
+                                                self.text_box_layout.layout_rect.height)
+                visible_bottom = current_height_adjustment + self.text_box_layout.view_rect.height
+                # handle cursor moving above current scroll position
+                if cursor_y_pos_top < current_height_adjustment:
+                    new_start_percentage = cursor_y_pos_top / self.text_box_layout.layout_rect.height
+                    self.scroll_bar.set_scroll_from_start_percentage(new_start_percentage)
+                # handle cursor moving below visible area
+                if cursor_y_pos_bottom > visible_bottom:
+                    new_top = cursor_y_pos_bottom - self.text_box_layout.view_rect.height
+                    new_start_percentage = new_top / self.text_box_layout.layout_rect.height
+                    self.scroll_bar.set_scroll_from_start_percentage(new_start_percentage)
+
+            self.redraw_from_text_block()
+
         if self.scroll_bar is not None and self.scroll_bar.check_has_moved_recently():
             height_adjustment = int(self.scroll_bar.start_percentage *
                                     self.text_box_layout.layout_rect.height)
@@ -465,6 +553,10 @@ class UITextBox(UIElement, IUITextOwnerInterface):
 
         if self.full_rebuild_countdown > 0.0:
             self.full_rebuild_countdown -= time_delta
+
+    def _handle_cursor_visibility(self):
+        # do nothing in text box - this is for the text entry box
+        pass
 
     def on_fresh_drawable_shape_ready(self):
         """
@@ -597,6 +689,7 @@ class UITextBox(UIElement, IUITextOwnerInterface):
                                              text_direction=self.parser.default_style['direction'],
                                              editable=True)
         self.text_box_layout.set_cursor_colour(self.text_cursor_colour)
+        self.text_box_layout.selection_colour = self.selected_bg_colour
         self.parser.empty_layout_queue()
         if self.dynamic_height:
             self.text_box_layout.view_rect.height = self.text_box_layout.layout_rect.height
@@ -684,86 +777,245 @@ class UITextBox(UIElement, IUITextOwnerInterface):
         self.link_hover_chunks = []
         self.text_box_layout.add_chunks_to_hover_group(self.link_hover_chunks)
 
-    def process_event(self, event: pygame.event.Event) -> bool:
+    def _process_mouse_button_event(self, event: Event) -> bool:
         """
-        Deals with input events. In this case we just handle clicks on any links in the text.
+        Process a mouse button event.
 
-        :param event: A pygame event to check for a reaction to.
+        :param event: Event to process.
 
-        :return: Returns True if we consumed this event.
+        :return: True if we consumed the mouse event.
 
         """
         consumed_event = False
         should_redraw_from_layout = False
-        if event.type == pygame.MOUSEBUTTONDOWN and event.button == 1:
+        if event.type == MOUSEBUTTONDOWN and event.button == BUTTON_LEFT:
             scaled_mouse_pos = self.ui_manager.calculate_scaled_mouse_position(event.pos)
             if self.hover_point(scaled_mouse_pos[0], scaled_mouse_pos[1]):
-                consumed_event = True
-
                 if self.is_enabled:
-                    if self.scroll_bar is not None:
-                        text_block_full_height = self.text_box_layout.layout_rect.height
-                        height_adjustment = (self.scroll_bar.start_percentage *
-                                             text_block_full_height)
-                    else:
-                        height_adjustment = 0
-                    base_x = int(self.rect[0] + self.padding[0] + self.border_width +
-                                 self.shadow_width + self.rounded_corner_width_offsets[0])
-                    base_y = int(self.rect[1] + self.padding[1] + self.border_width +
-                                 self.shadow_width + self.rounded_corner_height_offsets[0] - height_adjustment)
-                    for chunk in self.link_hover_chunks:
-
-                        hover_rect = pygame.Rect((base_x + chunk.x,
-                                                  base_y + chunk.y),
-                                                 chunk.size)
-                        if hover_rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1]):
-                            consumed_event = True
-                            if not chunk.is_active:
-                                chunk.set_active()
-                                should_redraw_from_layout = True
-
-        if self.is_enabled and event.type == pygame.MOUSEBUTTONUP and event.button == 1:
-            if self.scroll_bar is not None:
-                height_adjustment = (self.scroll_bar.start_percentage *
-                                     self.text_box_layout.layout_rect.height)
-            else:
-                height_adjustment = 0
-            base_x = int(self.rect[0] + self.padding[0] + self.border_width +
-                         self.shadow_width + self.rounded_corner_width_offsets[0])
-            base_y = int(self.rect[1] + self.padding[1] + self.border_width +
-                         self.shadow_width + self.rounded_corner_height_offsets[0] - height_adjustment)
-            scaled_mouse_pos = self.ui_manager.calculate_scaled_mouse_position(event.pos)
-            for chunk in self.link_hover_chunks:
-
-                hover_rect = pygame.Rect((base_x + chunk.x,
-                                          base_y + chunk.y),
-                                         chunk.size)
-                if (hover_rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1]) and
-                        self.rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1])):
                     consumed_event = True
-                    if chunk.is_active:
+                    should_redraw_from_layout = self._handle_hyper_link_mouse_button_down(scaled_mouse_pos)
+                    if not should_redraw_from_layout:
+                        text_layout_space_pos = self._calculate_text_space_pos(scaled_mouse_pos)
+                        self.text_box_layout.set_cursor_from_click_pos(text_layout_space_pos)
+                        self.edit_position = self.text_box_layout.get_cursor_index()
+                        self.redraw_from_text_block()
 
-                        # old event - to be removed in 0.8.0
-                        event_data = {'user_type': OldType(UI_TEXT_BOX_LINK_CLICKED),
-                                      'link_target': chunk.href,
-                                      'ui_element': self,
-                                      'ui_object_id': self.most_specific_combined_id}
-                        pygame.event.post(pygame.event.Event(pygame.USEREVENT, event_data))
+                        double_clicking = False
+                        if self.double_click_timer < self.ui_manager.get_double_click_time():
+                            if self._calculate_double_click_word_selection():
+                                double_clicking = True
+                                self.cursor_has_moved_recently = True
 
-                        # new event
-                        event_data = {'link_target': chunk.href,
-                                      'ui_element': self,
-                                      'ui_object_id': self.most_specific_combined_id}
-                        pygame.event.post(pygame.event.Event(UI_TEXT_BOX_LINK_CLICKED, event_data))
+                        if not double_clicking:
+                            self.select_range = [self.edit_position, self.edit_position]
+                            self.cursor_has_moved_recently = True
+                            self.selection_in_progress = True
+                            self.double_click_timer = 0.0
 
-                if chunk.is_active:
-                    chunk.set_inactive()
-                    should_redraw_from_layout = True
+        if (event.type == MOUSEBUTTONUP and
+                event.button == BUTTON_LEFT and
+                self.selection_in_progress):
+            scaled_mouse_pos = self.ui_manager.calculate_scaled_mouse_position(event.pos)
+
+            if self.hover_point(scaled_mouse_pos[0], scaled_mouse_pos[1]):
+                if self.is_enabled:
+                    consumed_event = True
+                    should_redraw_from_layout = self._handle_hyper_link_mouse_button_down(scaled_mouse_pos)
+                    if not should_redraw_from_layout:
+                        text_layout_space_pos = self._calculate_text_space_pos(scaled_mouse_pos)
+                        self.text_box_layout.set_cursor_from_click_pos(text_layout_space_pos)
+                        new_edit_pos = self.text_box_layout.get_cursor_index()
+                        if new_edit_pos != self.edit_position:
+                            self.edit_position = new_edit_pos
+                            self.cursor_has_moved_recently = True
+                            self.select_range = [self.select_range[0], self.edit_position]
+                            self.redraw_from_text_block()
+            self.selection_in_progress = False
 
         if should_redraw_from_layout:
             self.redraw_from_text_block()
 
         return consumed_event
+
+    def process_event(self, event: Event) -> bool:
+        consumed_event = False
+        if self._process_mouse_button_event(event):
+            consumed_event = True
+        if self.is_enabled and self.is_focused and event.type == KEYDOWN:
+            if self._process_keyboard_shortcut_event(event):
+                consumed_event = True
+            elif self._process_edit_pos_move_key(event):
+                consumed_event = True
+        return consumed_event
+
+    def _process_keyboard_shortcut_event(self, event: Event) -> bool:
+        """
+        Check if event is one of the CTRL key keyboard shortcuts.
+
+        :param event: event to process.
+
+        :return: True if event consumed.
+
+        """
+        consumed_event = False
+        if self._process_select_all_event(event):
+            consumed_event = True
+        elif self._process_copy_event(event):
+            consumed_event = True
+        return consumed_event
+
+    def _process_select_all_event(self, event: Event) -> bool:
+        """
+        Process a select all shortcut event. (CTRL+ A)
+
+        :param event: The event to process.
+
+        :return: True if the event is consumed.
+
+        """
+        consumed_event = False
+        if (event.key == K_a and
+                (event.mod & KMOD_CTRL or event.mod & KMOD_META) and
+                not (event.mod & KMOD_ALT)):  # hopefully enable diacritic letters
+            self._do_select_all()
+            consumed_event = True
+        return consumed_event
+
+    def _do_select_all(self):
+        if self.text_box_layout is not None:
+            self.select_range = [0, len(self.text_box_layout.plain_text)]
+            self.edit_position = len(self.text_box_layout.plain_text)
+            self.cursor_has_moved_recently = True
+
+    def _process_copy_event(self, event: Event) -> bool:
+        """
+        Process a copy shortcut event. (CTRL+ C)
+
+        :param event: The event to process.
+
+        :return: True if the event is consumed.
+
+        """
+        consumed_event = False
+        if (event.key == K_c and
+                (event.mod & KMOD_CTRL or event.mod & KMOD_META) and
+                not (event.mod & KMOD_ALT)):
+            self._do_copy()
+            consumed_event = True
+        return consumed_event
+
+    def _do_copy(self):
+        if self.text_box_layout is not None and abs(self.select_range[0] - self.select_range[1]) > 0:
+            low_end = min(self.select_range[0], self.select_range[1])
+            high_end = max(self.select_range[0], self.select_range[1])
+            clipboard_copy(self.text_box_layout.plain_text[low_end:high_end])
+
+    def _calculate_double_click_word_selection(self):
+        """
+        If we double-clicked on a word in the text, select that word.
+
+        """
+        if self.edit_position != self.select_range[0] or self.text_box_layout is None:
+            return False
+        index = min(self.edit_position, len(self.text_box_layout.plain_text) - 1)
+        if index >= 0:
+            char = self.text_box_layout.plain_text[index]
+            # Check we clicked in the same place on our second click.
+            pattern = re.compile(r"[\w']+")
+
+            while index >= 0 and not pattern.match(char):
+                index -= 1
+                if index >= 0:
+                    char = self.text_box_layout.plain_text[index]
+                else:
+                    break
+            while index >= 0 and pattern.match(char):
+                index -= 1
+                if index >= 0:
+                    char = self.text_box_layout.plain_text[index]
+                else:
+                    break
+            start_select_index = index + 1
+            index += 1
+            if index < len(self.text_box_layout.plain_text):
+                char = self.text_box_layout.plain_text[index]
+                while index < len(self.text_box_layout.plain_text) and pattern.match(char):
+                    index += 1
+                    if index < len(self.text_box_layout.plain_text):
+                        char = self.text_box_layout.plain_text[index]
+            end_select_index = index
+            self.select_range = [start_select_index, end_select_index]
+            self.edit_position = end_select_index
+            self.selection_in_progress = False
+            return True
+        else:
+            return False
+
+    def _calculate_text_space_pos(self, scaled_mouse_pos):
+        height_adjustment = 0
+        if self.scroll_bar is not None:
+            height_adjustment = (self.scroll_bar.start_percentage * self.text_box_layout.layout_rect.height)
+
+        text_layout_top_left = self.get_text_layout_top_left()
+        text_layout_space_pos = (scaled_mouse_pos[0] - text_layout_top_left[0],
+                                 scaled_mouse_pos[1] - text_layout_top_left[1] + height_adjustment)
+        return text_layout_space_pos
+
+    def _calculate_hyperlinks_offsets(self):
+        height_adjustment = 0
+        if self.scroll_bar is not None:
+            height_adjustment = (self.scroll_bar.start_percentage * self.text_box_layout.layout_rect.height)
+
+        text_layout_top_left = self.get_text_layout_top_left()
+        base_x = int(text_layout_top_left[0])
+        base_y = int(text_layout_top_left[1] - height_adjustment)
+
+        return base_x, base_y
+
+    def _handle_hyper_link_mouse_button_down(self, scaled_mouse_pos: Tuple[int, int]) -> bool:
+        should_redraw_from_layout = False
+        base_x, base_y = self._calculate_hyperlinks_offsets()
+        for chunk in self.link_hover_chunks:
+
+            hover_rect = pygame.Rect((base_x + chunk.x,
+                                      base_y + chunk.y),
+                                     chunk.size)
+            if hover_rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1]):
+                if not chunk.is_active:
+                    chunk.set_active()
+                    should_redraw_from_layout = True
+        return should_redraw_from_layout
+
+    def _handle_hyper_link_mouse_button_up(self, scaled_mouse_pos: Tuple[int, int]) -> bool:
+        should_redraw_from_layout = False
+        base_x, base_y = self._calculate_hyperlinks_offsets()
+
+        for chunk in self.link_hover_chunks:
+
+            hover_rect = pygame.Rect((base_x + chunk.x,
+                                      base_y + chunk.y),
+                                     chunk.size)
+            if (hover_rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1]) and
+                    self.rect.collidepoint(scaled_mouse_pos[0], scaled_mouse_pos[1])):
+
+                if chunk.is_active:
+                    # old event - to be removed in 0.8.0
+                    event_data = {'user_type': OldType(UI_TEXT_BOX_LINK_CLICKED),
+                                  'link_target': chunk.href,
+                                  'ui_element': self,
+                                  'ui_object_id': self.most_specific_combined_id}
+                    pygame.event.post(pygame.event.Event(pygame.USEREVENT, event_data))
+
+                    # new event
+                    event_data = {'link_target': chunk.href,
+                                  'ui_element': self,
+                                  'ui_object_id': self.most_specific_combined_id}
+                    pygame.event.post(pygame.event.Event(UI_TEXT_BOX_LINK_CLICKED, event_data))
+
+            if chunk.is_active:
+                chunk.set_inactive()
+                should_redraw_from_layout = True
+        return should_redraw_from_layout
 
     def rebuild_from_changed_theme_data(self):
         """
@@ -810,6 +1062,18 @@ class UITextBox(UIElement, IUITextOwnerInterface):
                                                              self.combined_element_ids)
         if border_colour != self.border_colour:
             self.border_colour = border_colour
+            has_any_changed = True
+
+        text_cursor_colour = self.ui_theme.get_colour_or_gradient('text_cursor',
+                                                                  self.combined_element_ids)
+        if text_cursor_colour != self.text_cursor_colour:
+            self.text_cursor_colour = text_cursor_colour
+            has_any_changed = True
+
+        selected_bg_colour = self.ui_theme.get_colour_or_gradient('selected_bg',
+                                                                  self.combined_element_ids)
+        if selected_bg_colour != self.selected_bg_colour:
+            self.selected_bg_colour = selected_bg_colour
             has_any_changed = True
 
         if self._check_text_alignment_theming():
@@ -1186,3 +1450,326 @@ class UITextBox(UIElement, IUITextOwnerInterface):
             pre_parsed_text = pre_parsed_text.replace('\n', '<br>')
 
         return pre_parsed_text
+
+    def unfocus(self):
+        """
+        Called when this element is no longer the current focus.
+        """
+        super().unfocus()
+
+    def focus(self):
+        """
+        Called when we 'select focus' on this element.
+        """
+        super().focus()
+        self.cursor_has_moved_recently = True
+
+    def _process_edit_pos_move_key(self, event: Event) -> bool:
+        """
+        Process an action key that is moving the cursor edit position.
+
+        :param event: The event to process.
+
+        :return: True if event is consumed.
+
+        """
+        consumed_event = False
+        # 4 left/right cases:
+        # 1. jump left or right one character (LEFT or RIGHT on their own)
+        # 2. jump to end/start of word (CTRL + LEFT or RIGHT)
+        # 3. jump to end/start of line (HOME or END)
+        # 4. jump to end/start of document (CTRL + HOME or END)
+        # plus 1 up or down a line/text row case (UP & DOWN arrows, CTRL makes no difference here in notepad,
+        # but in IDE it scrolls if possible)
+        # All cases can have selection attached with shift held
+
+        if event.key == K_LEFT:
+            if event.mod & KMOD_CTRL or event.mod & KMOD_META:
+                self._jump_edit_pos_to_start_of_word(should_select=(event.mod & KMOD_SHIFT))
+            else:
+                self._jump_edit_pos_one_character_left(should_select=(event.mod & KMOD_SHIFT))
+            consumed_event = True
+        elif event.key == K_RIGHT:
+            if event.mod & KMOD_CTRL or event.mod & KMOD_META:
+                self._jump_edit_pos_to_end_of_word(should_select=(event.mod & KMOD_SHIFT))
+            else:
+                self._jump_edit_pos_one_character_right(should_select=(event.mod & KMOD_SHIFT))
+            consumed_event = True
+        elif event.key == K_UP:
+            self._jump_edit_pos_one_row_up(should_select=(event.mod & KMOD_SHIFT))
+            consumed_event = True
+        elif event.key == K_DOWN:
+            self._jump_edit_pos_one_row_down(should_select=(event.mod & KMOD_SHIFT))
+            consumed_event = True
+        elif event.key == K_HOME:
+            if event.mod & KMOD_CTRL or event.mod & KMOD_META:
+                self._jump_edit_pos_to_start_of_all_text(should_select=(event.mod & KMOD_SHIFT))
+            else:
+                self._jump_edit_pos_to_start_of_line(should_select=(event.mod & KMOD_SHIFT))
+            consumed_event = True
+        elif event.key == K_END:
+            if event.mod & KMOD_CTRL or event.mod & KMOD_META:
+                self._jump_edit_pos_to_end_of_all_text(should_select=(event.mod & KMOD_SHIFT))
+            else:
+                self._jump_edit_pos_to_end_of_line(should_select=(event.mod & KMOD_SHIFT))
+            consumed_event = True
+        return consumed_event
+
+    def _jump_edit_pos_one_character_right(self, should_select=False):
+        if self.text_box_layout is not None:
+            if should_select:
+                if abs(self.select_range[0] - self.select_range[1]) > 0:
+                    if self.edit_position == self.select_range[1]:
+                        # try to extend to the right
+                        self.select_range = [self.select_range[0],
+                                             min(len(self.text_box_layout.plain_text), self.edit_position + 1)]
+                    elif self.edit_position == self.select_range[0]:
+                        # reduce to the right
+                        self.select_range = [min(len(self.text_box_layout.plain_text), self.edit_position + 1),
+                                             self.select_range[1]]
+                else:
+                    self.select_range = [self.edit_position, min(len(self.text_box_layout.plain_text),
+                                                                 self.edit_position + 1)]
+                if self.edit_position < len(self.text_box_layout.plain_text):
+                    self.edit_position += 1
+                    self.cursor_has_moved_recently = True
+            else:
+                if abs(self.select_range[0] - self.select_range[1]) > 0:
+                    self.edit_position = max(self.select_range[0], self.select_range[1])
+                    self.select_range = [0, 0]
+                    self.cursor_has_moved_recently = True
+                elif self.edit_position < len(self.text_box_layout.plain_text):
+                    self.edit_position += 1
+                    self.cursor_has_moved_recently = True
+
+    def _jump_edit_pos_one_character_left(self, should_select=False):
+        if should_select:
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                # existing selection, so edit_position should correspond to one
+                # end of it
+                if self.edit_position == self.select_range[0]:
+                    # try to extend to the left
+                    self.select_range = [max(0, self.edit_position - 1),
+                                         self.select_range[1]]
+                elif self.edit_position == self.select_range[1]:
+                    # reduce to the left
+                    self.select_range = [self.select_range[0],
+                                         max(0, self.edit_position - 1)]
+            else:
+                self.select_range = [max(0, self.edit_position - 1),
+                                     self.edit_position]
+            if self.edit_position > 0:
+                self.edit_position -= 1
+                self.cursor_has_moved_recently = True
+        else:
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                self.edit_position = min(self.select_range[0], self.select_range[1])
+                self.select_range = [0, 0]
+                self.cursor_has_moved_recently = True
+            elif self.edit_position > 0:
+                self.edit_position -= 1
+                self.cursor_has_moved_recently = True
+
+    def _jump_edit_pos_one_row_down(self, should_select=False):
+        if self.text_box_layout is not None:
+            if should_select:
+                new_pos = self.text_box_layout.get_cursor_pos_move_down_one_row(self.last_horiz_cursor_index)
+                if abs(self.select_range[0] - self.select_range[1]) > 0:
+                    if self.edit_position == self.select_range[1]:
+                        # try to extend to the right
+                        self.select_range = [self.select_range[0],
+                                             min(len(self.text_box_layout.plain_text), new_pos)]
+                    elif self.edit_position == self.select_range[0]:
+                        # reduce to the right
+                        self.select_range = [min(len(self.text_box_layout.plain_text), new_pos),
+                                             self.select_range[1]]
+                else:
+                    self.select_range = [self.edit_position, min(len(self.text_box_layout.plain_text),
+                                                                 new_pos)]
+                if self.edit_position < len(self.text_box_layout.plain_text):
+                    self.edit_position = new_pos
+                    self.cursor_has_moved_recently = True
+            else:
+                if abs(self.select_range[0] - self.select_range[1]) > 0:
+                    self.edit_position = self.text_box_layout.get_cursor_pos_move_down_one_row(self.last_horiz_cursor_index)
+                    self.select_range = [0, 0]
+                    self.cursor_has_moved_recently = True
+                else:
+                    self.edit_position = self.text_box_layout.get_cursor_pos_move_down_one_row(self.last_horiz_cursor_index)
+                    self.cursor_has_moved_recently = True
+                self.vertical_cursor_movement = True
+
+    def _jump_edit_pos_one_row_up(self, should_select=False):
+        if should_select:
+            new_pos = self.text_box_layout.get_cursor_pos_move_up_one_row(self.last_horiz_cursor_index)
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                # existing selection, so edit_position should correspond to one
+                # end of it
+                if self.edit_position == self.select_range[0]:
+                    # try to extend to the left
+                    self.select_range = [max(0, new_pos),
+                                         self.select_range[1]]
+                elif self.edit_position == self.select_range[1]:
+                    # reduce to the left
+                    self.select_range = [self.select_range[0],
+                                         max(0, new_pos)]
+            else:
+                self.select_range = [max(0, new_pos),
+                                     self.edit_position]
+            if self.edit_position > 0:
+                self.edit_position = new_pos
+                self.cursor_has_moved_recently = True
+        else:
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                self.edit_position = self.text_box_layout.get_cursor_pos_move_up_one_row(self.last_horiz_cursor_index)
+                self.select_range = [0, 0]
+                self.cursor_has_moved_recently = True
+            else:
+                self.edit_position = self.text_box_layout.get_cursor_pos_move_up_one_row(self.last_horiz_cursor_index)
+                self.cursor_has_moved_recently = True
+            self.vertical_cursor_movement = True
+
+    def _jump_edit_pos_to_start_of_word(self, should_select=False):
+        if self.text_box_layout is not None:
+            try:
+                text_to_search = self.text_box_layout.plain_text[max(0, self.edit_position - 1)::-1]
+                match_result = re.search(r'\b\w+\b', text_to_search)
+                if match_result is not None:
+                    next_pos = max(self.edit_position - match_result.regs[-1][1], 0)
+                else:
+                    next_pos = self.edit_position
+            except ValueError:
+                next_pos = 0
+            # include case when shift held down to select everything to start
+            # of current line.
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                if should_select:
+                    if self.edit_position == self.select_range[1]:
+                        # undo selection to right, create to left
+                        self.select_range = [next_pos, self.select_range[0]]
+                    else:
+                        # extend left
+                        self.select_range = [next_pos, self.select_range[1]]
+                else:
+                    next_pos = self.select_range[0]
+                    self.select_range = [0, 0]
+            else:
+                if should_select:
+                    self.select_range = [next_pos, self.edit_position]
+                else:
+                    self.select_range = [0, 0]
+            self.edit_position = next_pos
+            self.cursor_has_moved_recently = True
+
+    def _jump_edit_pos_to_end_of_word(self, should_select=False):
+        if self.text_box_layout is not None:
+            try:
+                match_result = re.search(r'\b\w+\b', self.text_box_layout.plain_text[self.edit_position:])
+                if match_result is not None:
+                    next_pos = min(self.edit_position + match_result.regs[0][1], len(self.text_box_layout.plain_text))
+                else:
+                    next_pos = self.edit_position
+            except ValueError:
+                next_pos = len(self.text_box_layout.plain_text)
+            # include case when shift held down to select everything to end
+            # of current line.
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                if should_select:
+                    if self.edit_position == self.select_range[0]:
+                        # undo selection to left, create to right
+                        self.select_range = [self.select_range[1], next_pos]
+                    else:
+                        # extend right
+                        self.select_range = [self.select_range[0], next_pos]
+                else:
+                    self.select_range = [0, 0]
+            else:
+                if should_select:
+                    self.select_range = [self.edit_position, next_pos]
+                else:
+                    self.select_range = [0, 0]
+            self.edit_position = next_pos
+            self.cursor_has_moved_recently = True
+
+    def _jump_edit_pos_to_start_of_line(self, should_select=False):
+        try:
+            next_pos = self.text_box_layout.set_cursor_to_start_of_current_row()
+        except ValueError:
+            next_pos = 0
+        # include case when shift held down to select everything to start
+        # of current line.
+        if abs(self.select_range[0] - self.select_range[1]) > 0:
+            if should_select:
+                if self.edit_position == self.select_range[1]:
+                    # undo selection to right, create to left
+                    self.select_range = [next_pos, self.select_range[0]]
+                else:
+                    # extend left
+                    self.select_range = [next_pos, self.select_range[1]]
+            else:
+                self.select_range = [0, 0]
+        else:
+            if should_select:
+                self.select_range = [next_pos, self.edit_position]
+            else:
+                self.select_range = [0, 0]
+        self.edit_position = next_pos
+        self.cursor_has_moved_recently = True
+
+    def _jump_edit_pos_to_end_of_line(self, should_select=False):
+        if self.text_box_layout is not None:
+            try:
+                next_pos = self.text_box_layout.set_cursor_to_end_of_current_row()
+            except ValueError:
+                next_pos = len(self.text_box_layout.plain_text)
+            # include case when shift held down to select everything to end
+            # of current line.
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                if should_select:
+                    if self.edit_position == self.select_range[0]:
+                        # undo selection to left, create to right
+                        self.select_range = [self.select_range[1], next_pos]
+                    else:
+                        # extend right
+                        self.select_range = [self.select_range[0], next_pos]
+                else:
+                    self.select_range = [0, 0]
+            else:
+                if should_select:
+                    self.select_range = [self.edit_position, next_pos]
+                else:
+                    self.select_range = [0, 0]
+            self.edit_position = next_pos
+            self.cursor_has_moved_recently = True
+
+    def _jump_edit_pos_to_start_of_all_text(self, should_select=False):
+        if should_select:
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                if self.edit_position == self.select_range[0]:
+                    self.select_range = [0, self.select_range[1]]
+                else:
+                    self.select_range = [0, self.select_range[0]]
+            else:
+                self.select_range = [0, self.edit_position]
+        else:
+            if abs(self.select_range[0] - self.select_range[1]) > 0:
+                self.select_range = [0, 0]
+        self.edit_position = 0
+        self.cursor_has_moved_recently = True
+
+    def _jump_edit_pos_to_end_of_all_text(self, should_select=False):
+        if self.text_box_layout is not None:
+            end_of_all_pos = len(self.text_box_layout.plain_text)
+            if should_select:
+                if abs(self.select_range[0] - self.select_range[1]) > 0:
+                    if self.edit_position == self.select_range[0]:
+                        self.select_range = [self.select_range[1], end_of_all_pos]
+                    else:
+                        self.select_range = [self.select_range[0], end_of_all_pos]
+                else:
+                    self.select_range = [self.edit_position, end_of_all_pos]
+            else:
+                if abs(self.select_range[0] - self.select_range[1]) > 0:
+                    self.select_range = [0, 0]
+            self.edit_position = end_of_all_pos
+            self.cursor_has_moved_recently = True
